# cocotola-question/api/openapi-bundled.yaml
openapi: 3.0.1
info:
  title: cocotola-question
  description: cocotola-question
  version: '1.0'
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers: []
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /api/v1/workbook/{workbookId}/question:
    post:
      summary: Add a question
      deprecated: false
      description: Add a new question to a workbook
      operationId: addQuestion
      tags:
        - question
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddQuestionRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
          headers: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Workbook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    get:
      summary: List questions
      deprecated: false
      description: List all questions in a workbook
      operationId: listQuestions
      tags:
        - question
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListQuestionsResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Workbook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
  /api/v1/workbook/{workbookId}/question/{questionId}:
    get:
      summary: Get a question
      deprecated: false
      description: Get a specific question by ID
      operationId: getQuestion
      tags:
        - question
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Question not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    put:
      summary: Update a question
      deprecated: false
      description: Update an existing question
      operationId: updateQuestion
      tags:
        - question
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: Successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
          headers: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Question not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
    delete:
      summary: Delete a question
      deprecated: false
      description: Delete an existing question
      operationId: deleteQuestion
      tags:
        - question
      parameters:
        - name: workbookId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '404':
          description: Question not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
...
  /api/v1/workbook/shared:
    get:
      summary: List shared workbooks
      deprecated: false
      description: List workbook references shared with the authenticated user
      operationId: listShared
      tags:
        - sharing
      parameters: []
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSharedResponse'
          headers: {}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers: {}
      security:
        - BearerAuth: []
        - CookieAuth: []
...
components:
  schemas:
...
    AddQuestionRequest:
      type: object
      properties:
        questionType:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required
          pattern: ^.*$
        content:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,max=10000
          maxLength: 10000
          pattern: ^[\s\S]*$
        orderIndex:
          type: integer
          format: int32
          x-oapi-codegen-extra-tags:
            binding: gte=0
      required:
        - questionType
        - content
        - orderIndex
    UpdateQuestionRequest:
      type: object
      properties:
        content:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required,max=10000
          maxLength: 10000
          pattern: ^[\s\S]*$
        orderIndex:
          type: integer
          format: int32
          x-oapi-codegen-extra-tags:
            binding: gte=0
      required:
        - content
        - orderIndex
    QuestionResponse:
      type: object
      properties:
        questionId:
          type: string
          x-go-name: QuestionID
          pattern: ^.*$
        questionType:
          type: string
          pattern: ^.*$
        content:
          type: string
          pattern: ^[\s\S]*$
        orderIndex:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - questionId
        - questionType
        - content
        - orderIndex
        - createdAt
        - updatedAt
    ListQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/QuestionResponse'
      required:
        - questions
...
    SharedItemResponse:
      type: object
      properties:
        referenceId:
          type: string
          x-go-name: ReferenceID
          pattern: ^.*$
        workbookId:
          type: string
          x-go-name: WorkbookID
          pattern: ^.*$
        addedAt:
          type: string
          format: date-time
      required:
        - referenceId
        - workbookId
        - addedAt
    ListSharedResponse:
      type: object
      properties:
        references:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/SharedItemResponse'
      required:
        - references
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          pattern: ^.*$
        message:
          type: string
          pattern: ^.*$
      required:
        - code
        - message
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token